Move egg/snake game to separate egg.py file
Extract EggGame class from MenuSessionHandler into its own module.
Update handlers.py to import and use EggGame for the snake easter egg.
This keeps the code modular and easier to maintain.
